mm, page_alloc: count movable pages when stealing from pageblock
When stealing pages from pageblock of a different migratetype, we count
how many free pages were stolen, and change the pageblock's migratetype if
more than half of the pageblock was free.  This might be too conservative,
as there might be other pages that are not free, but were allocated with
the same migratetype as our allocation requested.

While we cannot determine the migratetype of allocated pages precisely (at
least without the page_owner functionality enabled), we can count pages
that compaction would try to isolate for migration - those are either on
LRU or __PageMovable().  The rest can be assumed to be MIGRATE_RECLAIMABLE
or MIGRATE_UNMOVABLE, which we cannot easily distinguish.  This counting
can be done as part of free page stealing with little additional overhead.

The page stealing code is changed so that it considers free pages plus
pages of the "good" migratetype for the decision whether to change
pageblock's migratetype.

The result should be more accurate migratetype of pageblocks wrt the
actual pages in the pageblocks, when stealing from semi-occupied
pageblocks.  This should help the efficiency of page grouping by mobility.

In testing based on 4.9 kernel with stress-highalloc from mmtests
configured for order-4 GFP_KERNEL allocations, this patch has reduced the
number of unmovable allocations falling back to movable pageblocks by 47%.
The number of movable allocations falling back to other pageblocks are
increased by 55%, but these events don't cause permanent fragmentation, so
the tradeoff should be positive.  Later patches also offset the movable
fallback increase to some extent.

Link: http://lkml.kernel.org/r/20170307131545.28577-5-vbabka@suse.cz
Signed-off-by: Vlastimil Babka <vbabka@suse.cz>
Acked-by: Mel Gorman <mgorman@techsingularity.net>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Cc: Joonsoo Kim <iamjoonsoo.kim@lge.com>
Cc: David Rientjes <rientjes@google.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>