mm: remove seemingly spurious reclaimability check from laptop_mode gating
1d82de618ddd ("mm, vmscan: make kswapd reclaim in terms of nodes") allowed
laptop_mode=1 to start writing not just when the priority drops to
DEF_PRIORITY - 2 but also when the node is unreclaimable.  That appears to
be a spurious change in this patch as I doubt the series was tested with
laptop_mode, and neither is that particular change mentioned in the
changelog.  Remove it, it's still recent.

Link: http://lkml.kernel.org/r/20170228214007.5621-4-hannes@cmpxchg.org
Signed-off-by: Johannes Weiner <hannes@cmpxchg.org>
Acked-by: Hillf Danton <hillf.zj@alibaba-inc.com>
Acked-by: Mel Gorman <mgorman@techsingularity.net>
Acked-by: Michal Hocko <mhocko@suse.com>
Cc: Jia He <hejianet@gmail.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>