HACK: CHROMIUM: cpufreq: sched: Don't change frequencies when stopped
The cpufreq_sched_stop() (called as a result of the CPUFREQ_GOV_STOP
selector) wasn't actually causing the "sched" governor to stop
changing frequencies.  Specifically the cpufreq_sched_thread() kept
running after the stop happened and that could keep changing
frequencies.

The lack of stoppiness in the sched governor is a problem when we have
a board relying on "opp-suspend" support.  One such board is
rk3399-kevin, which relies on the cpu frequency to be less than a
certain speed during suspend.  In rk3399-kevin's case this is because
the suspend process causes certain regulators to drop down to a
voltage that cannot support the highest CPU speed.

The specific order of calls that shows how this is supposed to work
(at least on rk3399-kevin, which uses cpufreq-dt) is:

1. cpufreq_suspend() calls __cpufreq_governor(..., CPUFREQ_GOV_STOP)
2. cpufreq_suspend() calls cpufreq_driver->suspend(...)
3. cpufreq_driver->suspend is cpufreq_generic_suspend()
4. cpufreq_generic_suspend() sets things to the suspend frequency.

...as you can see, if the CPUFREQ_GOV_STOP doesn't actually stop the
governor then it's possible we'll end up at a frequency other than the
suspend frequency.  If it's higher than the suspend frequency we might
go boom.

A few notes about this implementation:

A) This is a "quick and dirty" implementation.  I don't actually
   change any governor logic--I just stop the frequency from changing
   at times when it shouldn't.  Presumably this could mess up the
   cpufreq accounting slightly if the driver thinks it's running at a
   different rate than it really is.  Hopefully any effects should be
   short term.
B) I don't rely on the "policy->rwsem" for protecting the new
   "started" member because it appears that cpufreq is not consistent
   about grabbing that semaphore before calling CPUFREQ_GOV_STOP.
   Ugh.  For old times take, I continue to grab "policy->rwsem" in
   cpufreq_sched_try_driver_target() where it used to claim "avoid
   race with cpufreq_sched_stop" even though it's unclear how/why that
   was supposed to work.  It seems safer to leave the semaphore grab
   in case it had some purpose.

BUG=chrome-os-partner:61705
TEST=suspend_stress_test w/ sched governor
TEST=Put in test printouts; see cases where we avoided problems

Change-Id: I634e0122a10482f44b70d683cf10f58be8306c58
Signed-off-by: Douglas Anderson <dianders@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/426010
Reviewed-by: Sonny Rao <sonnyrao@chromium.org>

Conflicts:
	kernel/sched/cpufreq_sched.c

[rebase49(groeck): Resolved conflicts]
Signed-off-by: Guenter Roeck <groeck@chromium.org>