ARM: 8674/1: dma-mapping: Reset the device's dma_ops
arch_teardown_dma_ops() being the inverse of arch_setup_dma_ops(),
dma_ops should be cleared in the teardown path. Otherwise this causes
problem when the probe of device is retried after being deferred. The
device's iommu structures are cleared after EPROBEDEFER error, but on
the next try dma_ops will still be set to old value, which is not right.

Fixes: 26b37b946a5c ("arm: dma-mapping: Don't override dma_ops in arch_setup_dma_ops()")
Signed-off-by: Sricharan R <sricharan@codeaurora.org>
Reviewed-by: Robin Murphy <robin.murphy@arm.com>
Acked-by: Marek Szyprowski <m.szyprowski@samsung.com>
Signed-off-by: Russell King <rmk+kernel@armlinux.org.uk>