target: Fix overflow/underflow handling of commands with a Data-Out buffer
As required by the SCSI specs, report READ and WRITE commands
with a Data-Out buffer that has another size than what is
specified in the CDB as overflow or underflow instead of
"invalid field in CDB". Although this patch does not change the
behavior for regular SCSI initiators, this patch is helpful
because it reduces the number of compliance failures reported by
the libiscsi test tool.

Signed-off-by: Bart Van Assche <bart.vanassche@sandisk.com>
Cc: Christoph Hellwig <hch@lst.de>
Cc: Hannes Reinecke <hare@suse.com>
Cc: David Disseldorp <ddiss@suse.de>