mm: drop HASH_ADAPT
"mm: Adaptive hash table scaling" has introduced a new large hash table
automatic scaling because the previous implementation led to too large
hashes on TB systems.  This is all nice and good but the patch assumes
that callers of alloc_large_system_hash will opt-in to use this new
scaling.  This makes the API unnecessarily complicated and error prone.
The only thing that callers should care about is whether they have an
upper bound for the size or leave it to alloc_large_system_hash to decide
(by providing high_limit == 0).

As a quick code inspection shows there are users with high_limit == 0
which do not use the flag already e.g.  {dcache,inode}_init_early or
mnt_init when creating mnt has tables.  They certainly have no good reason
to use a different scaling because the [di]cache was the motivation for
introducing a different scaling in the first place (we just do this
attempt and use memblock).  It is also hard to imagine why we would mnt
hash tables need larger hash tables.

Just drop the flag and use the scaling whenever there is no high_limit
specified.

Link: http://lkml.kernel.org/r/20170509094607.GG6481@dhcp22.suse.cz
Signed-off-by: Michal Hocko <mhocko@suse.com>
Reviewed-by: Pavel Tatashin <pasha.tatashin@oracle.com>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>