msm: sps: Optimize for device handling by single core
When the crypto device is handled by single CPU using polling,
crypto driver guarantees the exclusive access to the device so
that the spin lock can be avoided. The polling thread does the
EOT polling before issuing any new request, there is no need to
check EOT when issuing the new request. To avoid the AHB access,
the read point is cached.

Change-Id: Ie54be9594fa266c3425a54073eaf0301d29ca460
Acked-by: Jie Luo <jluo@qti.qualcomm.com>
CRs-Fixed: 2003560
Signed-off-by: Yasir Malik <ymalik@codeaurora.org>