mm: make kswapd try harder to keep active pages in cache
When testing a slab heavy workload I noticed that we often would barely
reclaim anything at all from slab when kswapd started doing reclaim.  This
is because we use the ratio of nr_scanned / nr_lru to determine how much
of slab we should reclaim.  But in a slab only/mostly workload we will not
have much page cache to reclaim, and thus our ratio will be really low and
not at all related to where the memory on the system is.  Instead we want
to use a ratio of the reclaimable slab to the actual reclaimable space on
the system.  That way if we are slab heavy we work harder to reclaim slab.

The other part of this that hurts is when we are running close to full
memory with our working set.  If we start putting a lot of reclaimable
slab pressure on the system (think find /, or some other silliness), we
will happily evict the active pages over the slab cache.  This is kind of
backwards as we want to do all that we can to keep the active working set
in memory, and instead evict these short lived objects.  The same thing
occurs when say you do a yum update of a few packages while your working
set takes up most of RAM, you end up with inactive lists being relatively
small and so we reclaim active pages even though we could reclaim these
short lived inactive pages.

My approach here is twofold.  First, keep track of the difference in
inactive and slab pages since the last time kswapd ran.  In the first run
this will just be the overall counts of inactive and slab, but for each
subsequent run we'll have a good idea of where the memory pressure is
coming from.  Then we use this information to put pressure on either the
inactive lists or the slab caches, depending on where the pressure is
coming from.

If this optimization does not work, then we fall back to the previous
methods of reclaiming space with a slight adjustment.  Instead of using
the overall scan rate of page cache to determine the scan rate for slab,
we instead use the total usage of slab compared to the reclaimable page
cache on the box.  This will allow us to put an appropriate amount of
pressure on the slab shrinkers if we are a mostly slab workload.

I have two tests I was using to watch either side of this problem.  The
first test kept 2 files that took up 3/4 of the memory, and then started
creating a bunch of empty files.  Without this patch we would have to
re-read both files in their entirety at least 3 times during the run.
With this patch the active pages are never evicted.

The second test was a test that would read and stat all the files in a
directory, which again would take up about 3/4 of the memory with slab
cache.  Then I cat'ed a 100gib file into /dev/null and checked to see if
any of the files were evicted and verified that none of the files were
evicted.

Link: http://lkml.kernel.org/r/1495549403-3719-1-git-send-email-jbacik@fb.com
Signed-off-by: Josef Bacik <jbacik@fb.com>
Acked-by: Rik van Riel <riel@redhat.com>
Cc: Johannes Weiner <hannes@cmpxchg.org>
Signed-off-by: Andrew Morton <akpm@linux-foundation.org>